Updated api_views.py
Updated the code to verify vendor and vendor of product being added, not just user with all authentication, but proper vendor that is registered. I have also added in checks to make sure it does not allow duplicates for store or product which I should have added in the first instance. Added #IMPROVE to these